fix(ci): resolve PR #217 CI blockers
PR #217 (Issue #194, feat(iot): manufacturer-signed firmware verification) had three primary CI failures plus several pre-existing workspace-wide clippy lints. This commit brings every validator green.

Primary CI blockers (PR #217 specific):
- contracts/iot_device_management: deleted broken soroban_sdk::contractimport!-based crypto_registry_client.rs shim (only accepts pre-built .wasm via CARGO_MANIFEST_DIR resolution, not Rust source). Added direct workspace path-dep on crypto_registry. Refactored register_device from 11 ABI params to 7 by introducing DeviceRegistrationInfo and FirmwareAttestation structs (Soroban hard 10-param limit). Rewrote verify_manufacturer_firmware_signature as a 4-arm match against the SDK nested Result<Result<Option<KeyBundle>, _>, _> return shape.
- Cargo.toml [profile.release]: lto = "true" -> lto = "thin" to dodge cross-crate bitcode corruption when iot_device_management (cdylib) consumes crypto_registry (cdylib+rlib) path dep.

Test infrastructure unblocking:
- contracts/healthcare_reputation/Cargo.toml: added rlib to [lib] crate-type so tests/ can use healthcare_reputation::*.
- tests/integration/: renamed local mod healthcare_reputation -> healthcare_reputation_tests to break the local-mod shadow of the external crate. Switched rejection assertions to matches! on the SDK nested Result<Result<T, E>, InvokeError>. Pinned test ledger timestamps to avoid now.saturating_sub(N) saturation to 0.
- libs/access_utils: #[allow(clippy::result_unit_err)] on init_admin (Result<(), ()> is workspace-wide contract initialization signal).

Pre-existing workspace clippy lints surfaced by cold sweep:
- contracts/identity_registry: dropped unused let-binding on test fixture.
- contracts/medical_records: dropped redundant `use access_utils;` (single-component path import under Rust 2018+).
- contracts/storage_cleanup: dropped unnecessary u32->u32 cast on kept_indices.len().
- contracts/zkp_registry: targeted #[allow(dead_code)] on TD-001/TD-003 scaffolding helpers.

Docs regenerated:
- docs/API_REFERENCE.md, docs/PUBLIC_API.md, docs/portal/index.html, schemas/docs/api.json, docs/ERROR_CODES.md.

Verified locally:
  cargo fmt -- --check                                                 clean
  cargo clippy --workspace --all-targets -- -D warnings                clean
  cargo test --workspace                                               exit 0
  cargo build --target wasm32-unknown-unknown --release per-contract   all 57 clean
  docs regen + git diff --quiet                                        clean

Closes #217

+++ contracts/storage_cleanup/src/lib.rs
@@ -502,7 +502,7 @@ impl StorageCleanup {
         }
 
         // Update count
-        let new_count = kept_indices.len() as u32;
+        let new_count = kept_indices.len();
         env.storage()
             .persistent()
             .set(&DataKey::CategoryCount(category), &new_count);
